samples/kubernetes/src/pipeline-server-k8s-controller/requirements.txt
paho-mqtt==1.6.1
requests==2.28.1
tornado==6.2
kubernetes==24.2.0
